Actualizando todo con la nueva base de datos - Ambrosio

+++ registerSpringBoot/src/main/java/com/proyecto/encuentranos/repositorios/IServicioRepositorio.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.proyecto.encuentranos.modelos.ServicioModelo;
 //CREANDO EL REPOSITORIO PARA ServicioModelo
 @Repository
-public interface IServicioRepositorio extends JpaRepository<ServicioModelo, Long>{
+public interface IServicioRepositorio extends JpaRepository<ServicioModelo, Integer>{
 
 }
